% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{check_contiguity}
\alias{check_contiguity}
\title{Check contiguity of all clusters}
\usage{
check_contiguity(sf_data, adj_matrix)
}
\arguments{
\item{sf_data}{An \code{sf} object with a \code{cluster} column.}

\item{adj_matrix}{Adjacency matrix from \code{\link[=prepare_districts]{prepare_districts()}}.}
}
\value{
Named logical vector (one element per cluster).
}
\description{
Returns a named logical vector indicating whether each cluster in
\code{sf_data$cluster} forms a geographically contiguous region.
}
