% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegSwarmEP-package.R
\docType{package}
\name{SegSwarmEP-package}
\alias{SegSwarmEP}
\alias{SegSwarmEP-package}
\title{SegSwarmEP: Ant Colony Optimization for School District Redistricting to Reduce Segregation}
\description{
Implements an Ant Colony Optimization (ACO) algorithm for redistricting school districts to minimize racial and demographic segregation, as measured by Theil's H information index. The algorithm respects geographic contiguity constraints, optimizes compactness (Polsby-Popper score), and uses adaptive elite pheromone scheduling. Designed to work with any state or regional dataset provided as an 'sf' spatial data frame, making it applicable nationwide for school desegregation planning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wootenjp/SegSwarmEP-R-}
  \item Report bugs at \url{https://github.com/wootenjp/SegSwarmEP-R-/issues}
}

}
\author{
\strong{Maintainer}: J. P. Wooten \email{wootenjp@vcu.edu}

}
\keyword{internal}
